refactor project structure
- move header related modules into `header` submodule
- rename `raw.rs` to `message_raw_crypto` as it is related to `Message`
  struct and `raw-crypto` feature
- move helper functions from `message.rs` into `helpers`
  sub(sub)modules, grouped by purpose:
    - encryption and key handling
    - receiving messages
    - serialization of data
- move some tests from `message.rs` into `/tests` folder
